I made a new skin from BN6 and made sure all the measurements were perfect. I also rewrote the skin parser to be a bit more functional (now people can comment anywhere in skin file, put values in any order, etc.) and to be easy to add functionality to in the future.
Coming next: Maps and backgrounds! New problem: Some maps have transparent pieces (and multiple pieces for that matter) so we can't do the transparent thing like before. There are two solutions I can think of: collision maps (black and white image, one color is where you can walk) or make the map tiled (way harder initially, but could be easier to animate).

One thing I want to know how to do efficiently is to animate small pieces of the map. That is not have an image of the entire map for each frame. Any idea how this would work?